requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
google-generativeai==0.8.3
python-dotenv==1.0.1